pyproject.toml
[tool.poetry]
name = "delivery-fee-calculator"
version = "0.1.0"
description = ""
authors = ["Haowei Lee <haoweilee.tpe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.81.0"
uvicorn = "^0.18.3"
gunicorn = "^20.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "7.1.3"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
